refactor dfp to gam (#13192)
