+++ integrationExamples/gpt/adloox.html
@@ -117,7 +117,7 @@
                 if (videoBids) {
 // DEMO NOTES: your environment likely will use the commented section ////
                     var videoUrl = videoBids.bids[0].vastUrl;
-//                    var videoUrl = pbjs.adServers.dfp.buildVideoUrl({
+//                    var videoUrl = pbjs.adServers.gam.buildVideoUrl({
 //                        adUnit: videoAdUnit,
 //                        params: {
 //                            iu: '/19968336/prebid_cache_video_adunit',
