+++ integrationExamples/gpt/localCacheGam.html
@@ -95,7 +95,7 @@
 
             const bid = bidResponse.bids[0];
 
-            const vastXml = await pbjs.adServers.dfp.getVastXml({
+            const vastXml = await pbjs.adServers.gam.getVastXml({
               bid,
               adUnit: 'div-gpt-ad-51545-0',
               params: {
