+++ libraries/gamUtils/gamUtils.js
@@ -1,0 +1,1 @@
+export {DEFAULT_DFP_PARAMS as DEFAULT_GAM_PARAMS, DFP_ENDPOINT as GAM_ENDPOINT, gdprParams} from '../dfpUtils/dfpUtils.js';
