+++ modules/.submodules.json
@@ -59,7 +59,7 @@
     ],
     "adpod": [
       "freeWheelAdserverVideo",
-      "dfpAdpod"
+      "gamAdpod"
     ],
     "rtdModule": [
       "1plusXRtdProvider",
