+++ modules/ixBidAdapter.md
@@ -393,10 +393,10 @@ var adUnits = [{
 
 ### 2. Include `ixBidAdapter` in your build process
 
-When running the build command, include `ixBidAdapter` as a module, as well as `dfpAdServerVideo` if you require video support.
+When running the build command, include `ixBidAdapter` as a module, as well as `gamAdServerVideo` if you require video support.
 
 ```
-gulp build --modules=ixBidAdapter,dfpAdServerVideo,fooBidAdapter,bazBidAdapter
+gulp build --modules=ixBidAdapter,gamAdServerVideo,fooBidAdapter,bazBidAdapter
 ```
 
 If a JSON file is being used to specify the bidder modules, add `"ixBidAdapter"`
@@ -405,7 +405,7 @@ to the top-level array in that file.
 ```json
 [
     "ixBidAdapter",
-    "dfpAdServerVideo",
+    "gamAdServerVideo",
     "fooBidAdapter",
     "bazBidAdapter"
 ]
