+++ modules/videoModule/gamAdServerSubmodule.js
@@ -3,10 +3,10 @@ import { getGlobal } from '../../src/prebidGlobal.js';
 
 /**
  * @class
- * @param {Object} dfpModule_ - the DFP ad server module
+ * @param {Object} gamModule_ - the GAM ad server module
  */
-function GamAdServerProvider(dfpModule_) {
-  const dfp = dfpModule_;
+function GamAdServerProvider(gamModule_) {
+  const dfp = gamModule_;
 
   function getAdTagUrl(adUnit, baseAdTag, params, bid) {
     return dfp.buildVideoUrl({ adUnit: adUnit, url: baseAdTag, params, bid });
@@ -23,7 +23,7 @@ function GamAdServerProvider(dfpModule_) {
 }
 
 export function gamSubmoduleFactory() {
-  const dfp = getGlobal().adServers.dfp;
+  const dfp = getGlobal().adServers.gam;
   const gamProvider = GamAdServerProvider(dfp);
   return gamProvider;
 }
