+++ test/pages/instream.html
@@ -16,7 +16,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/videojs-vast-vpaid/2.0.2/videojs_5.vast.vpaid.min.js"></script>
 
     <!-- prebid.js -->
-    <script src="http://localhost:4444/bundle?modules=appnexusBidAdapter&modules=dfpAdServerVideo" async=true></script>
+    <script src="http://localhost:4444/bundle?modules=appnexusBidAdapter&modules=gamAdServerVideo" async=true></script>
 
 
     <script>
@@ -69,7 +69,7 @@
         pbjs.requestBids({
           timeout: 10000,
           bidsBackHandler : function(bids) {
-            var videoUrl = pbjs.adServers.dfp.buildVideoUrl({
+            var videoUrl = pbjs.adServers.gam.buildVideoUrl({
               adUnit: videoAdUnit,
               params: {
                 iu: '/19968336/prebid_cache_video_adunit'
